vcl tiled rendering: avoid 'Pixel represents color values!' warnings
Which happens e.g. during rendering tiles:
OutputDevice::BlendBitmapWithAlpha() produces an RGB alpha
(255,255,255), then tries to set a pixel in the 8bit alpha channel using
it, so when BitmapReadAccess::SetPixelFor_8BIT_PAL() tries to get the
color index, we try to get it from a color that is not indexed.

Let's assume that when the color is not indexed, it's always gray, so it
doesn't matter what color we pick for the alpha mask needs.

Change-Id: I325c1d70514fd176fdc9cc39683b444447adf07f