sw: fix copying of bookmarks in CopyRange
If the copied range starts with a not fully selected paragraph, the
bookmarks that are copied will be created on the wrong paragraphs,
on the node after the correct one.

This also happens when hinding the redlines, and causes asserts from
attempting to create CrossRefBookmarks on table nodes on WW8 export of
fdo66302-1.odt and fdo66312-1.odt.

Change-Id: Id576be3e38a89527d967f02b39d9aabbf6368354