Allow non ASCII chars as accelerator part 2
This is a follow up patch to the commit:
d5cb25cd532fc2ba65c73f3dcc2f6d9ec8329d28
This takes care of exposing non ascii characters to atk so that the
orca screen reader has a chance to present these characters to it's users.
I was considering removing the KeyCode bits all toghether and rely on
what we get from KeyChar but got cold feet in the end.

Change-Id: I7a3e78c47b81f8b36763c56e0298e8bbf226dccf
Reviewed-on: https://gerrit.libreoffice.org/19105
Reviewed-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>
Tested-by: Samuel Mehrbrodt <Samuel.Mehrbrodt@cib.de>