crashtesting: remove old fix for ooo55381-1.sxw
this reverts

commit 97075cc94e6547a4078050f8fa9f0ce37638c6f9
Author: Caolán McNamara <caolanm@redhat.com>
Date:   Wed Jun 3 14:15:03 2015 +0100

    don't crash on layout of ooo55381-1.sxw

    The pLay here is getting joined and thus deleted in a stack of
    FormatLayoutTabs so set it as unjoinable for the duration that
    we expect pLay to remain valid

because ooo31216-2.doc fdo57780-2.doc rhbz193771-1.doc assert when converting
it to a more conservative SwFrmDeleteGuard. They would leave the function
safely on the early return if the delete was allowed, so go more conservative
again and restore to the original logic.

ooo55381-1.sxw continues to apparently not fail.