vcl: recolor images (icons) to be more visible in a dark theme
This adds recoloring of images/icons (suited for breeze and sifr)
so they are more visible when a dark theme is used (for example gtk3
can switch to a dark theme). LO must be started with environment
variable VCL_ICONS_FOR_DARK_THEME set to 1.

open issues - currently all images/icons are recolored but not all
images should get recolored.

Change-Id: Ibc42e30af79bb4c4f04c67b760019311b97b2cc6