tdf#89165 sw: looping layout in SwHeadFootFrm::FormatSize()
FormatObjsAtFrm() returns false, indicating that the result is not
perfect, and the caller hopes that formatting all children of the frame
again will help. Make sure that in case that assumption turns out to be false, we at
least do not loop.

The loop limit is enough to make the original i#43771 bugdoc still laid
out properly.

Change-Id: Ife9bd99a628a72edb922d04e05081971b6adf340