Automating configure choices based on what is installed considered harmful
Don't automatically --disable-odk (if not mentioned explicitly either
way) based on whether doxygen is found or not.

Caolán says: It's an absolute pain as a maintainer when packages do
that. You build it in some minimal build env and all is well, then
some depend changes and something else ends up in the build env and
now your package fails to build anymore, or behaves quite differently.

Change-Id: I8bc6ab6f90e6e070a37e37b5108081425e116173
Reviewed-on: https://gerrit.libreoffice.org/19324
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Michael Stahl <mstahl@redhat.com>