These places apparently want to unbox a value boxed as an Any
...but what the original code would have done is try to unbox as an Any again,
which would throw IllegalArgumentException.  As the unboxed value would only be
used for printing to System.out (and Any.toString result is just fine), just
don't bother to unbox at all.  (In the second place, the oldValue result would
further be used in a util.ValueComparer.equalValue call, but that internally
takes care of correctly unboxing its arguments, anyway.)

Change-Id: I6802d1acd787f19346f66b418372be1701f69139