sd search: restore start selection if nothing was found
For one, if sd::Outliner::RememberStartPosition() cares about storing
the current selection in the DrawViewShell case in maStartSelection,
then sd::Outliner::RestoreStartPosition() should probably restore it.

For another, sd::Outliner::StartSearchAndReplace() returned true in the
"nothing was found" case, which restarted the spellchecking, which
killed the selection. If the selection has been restored already, then
at least don't kill it in the find case: find is read-only, so no need
to restart the spellchecker.

Change-Id: Idd886d449a9cfb164887fc6b1fde27b5e63e272b