Writer image export: cmd. line, default to white background
This commit fixes the writer image export (jpeg, png) that didn't
work because the export pixel size was set to 0 by default. Now
the default is set to document size (which depends on system DPI).

When exporting to a PNG the background was transparent, which may
not  be desired. The background color is now by default white and
can be changed for DocumentToGraphicRenderer, but the dialog or
command line don't support such an option - for now.

Change-Id: I16ffd3cd60c47b52768f43ae4c4c170fc821033b