gbuild: incremental builds with --enable-pch are unsound
The problem is that the precompiled headers' dependency files are not
run through concat-deps, hence they directly refer to headers of
external libraries' headers, which are not targets in the build system;
therefore re-building an external library does not cause the dependent
PCH to re-build and (at least with MSVC) the object files don't depend
on headers included via PCH anyway, so we get the recent link failure
in comphelper with MSVC, which wasn't rebuilt for icu's ABI change.

To fix that just use concat-deps, which re-writes header dependencies to
UnpackedTarball target dependencies.

Change-Id: Ic7555822925aaa1ff09b29bb73801fb83923bfab