tdf#94138 fix printing of edit form fields
There are 2 bugs:

- In Window::GetDrawPixelFont we asked for the font to the
current window, which didn't work correctly when printing. The
device itself should provide the font and not the window.

- In paint method the font wasn't setup correctly which should
happen before painting by calling ApplySettings. The effect of
this is that print preview didn't show the correct font for the
edit control and similar when you copy + paste the edit control.
The work around for this is to call ApplySettings again.

Change-Id: I74960355823c71c1d5a18a82bbc86561a3c4b760