Make number recognition work in writer tables again
It seems that number recognition in tables are not working properly
enter 10-10-10 and it should be converted to a date but it is not.
I tracked it down to the fix of bug fdo#32082. It looks like bSetNumFmt
was changed to false by mistake. Since then it has changed name to
bSetNumFormat. From what I can tell fdo#32082 still works after this
patch, but I might have missed some nuance of that bug report.

Added two tests, one for the bug mentioned above and one to check
that number recognition is working. At least with a simple date.

Change-Id: Id58849a223eb602054c66c7379cd56a68a93dea2
Reviewed-on: https://gerrit.libreoffice.org/19563
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Jan Holesovsky <kendy@collabora.com>