Revert "tdf#90695: In function Wizard Array box cannot be unchecked"
This reverts commit 9bc2b34ed49eef7376df76191e67dc8c6d520e9a.

Removing the disabling is plain wrong. Please see the code for when the button is disabled, when it is enabled, and what consequences it has when changing the state from checked to unchecked. Unchecking will result in the formula not being applied as array formula anymore, which it was before editing.

Actually tdf#90695 is not even a bug. The checked but disabled Array button is meant as an indicator that the formula IS an array/matrix formula.

Change-Id: Ic7f9f7010ce208171ef614e0885474f2deae4ac9
Reviewed-on: https://gerrit.libreoffice.org/19602
Reviewed-by: Eike Rathke <erack@redhat.com>
Tested-by: Eike Rathke <erack@redhat.com>