tdf#94022 Print dialog: hide Installable Options
Options specified in PPD files can be grouped using the
keywords "OpenGroup" and "CloseGroup".
The keyword "InstallableOptions" is used as a group name
for a group containing options that define
optional hardware features of the printer that can
be present or not (s. section 5.4 in version 4.3 of
the PPD specification).

As they are not print job specific, it is recommended
not to show them in the print dialog.

To be able to distinguish those options, the
PPD group name was added as an attribute to the PPDKey
class.

Change-Id: I4a3abf23a711ad98556c0b608a07ef0a91e77e2b
Reviewed-on: https://gerrit.libreoffice.org/19623
Tested-by: Jenkins <ci@libreoffice.org>
Reviewed-by: Katarina Behrens <Katarina.Behrens@cib.de>