sw: fix "long" overflow in lcl_CalcWish
0 + (1 * 32768 * 65536 / 65537) evaluates to -32767 with 32-bit MSVC
whereas with 64-bit GCC the result is +32767.

The multiplication overflows to 0x80000000 - so do that in 64 bits.

This fixes the JunitTest_sw_unoapi_4 failure in SwXTextTable on the
TableBorder property, where the wrong result caused MakeSelUnions()
to believe that the last cell is not visible and hence setting the
TableBorder property failed.

Change-Id: Idb10535ce5342cd83d7ab6aea07782a2034469eb