sc lok: Cache viewdata zoom and reuse for cursor callback
As of 716be53d82c709b00c180bc197d2dbcd179571cd we reset the zoom
level to the default when processing LOK mouse events. The exact
cell cursor position is dependent on the zoom level (due to the
rounding in the cell position summing calculations), hence we need
to make sure we have the correct zoom level for those calculations
(or else the rounding errors will result in incorrect cell cursor
positions). Caching the zoom level and reusing it only here seems
to be the most efficient way of solving this for now.

(An alternative would be to only send the cell ID in the callback,
 and have the frontend then request the pixel positions together
 with the current frontend zoom level - however especially for
 LOOL minimising the number of trips is probably wise.)

Change-Id: Iae3aabfd7ea9bec7057be7b63670885766870c4f