Resolves: #i61277# Basic open file with random access erases all data.
StarBasic tries to emulate truncation of files opened for writing
by deleting them before opening them (if they exist). However there
are exclusion to the truncation behaviour when the file is opened
in append mode or is a binary file. Another exclusion - that was
missing with disasterous consequences - is that the file should
not be truncated when opened in random access mode.

(cherry picked from commit 72b2f07ccac7578ef36fc2b92dcba49abe397ebc)

Change-Id: Ic0053039ce5fd67ad42bf5332ae301d83f1a158d