sw lok: fix width of the notes sidebar
The map mode is in general disabled during tiled rendering, so mouse
positions can be sent in using twips, but here we have to temporarily
enable it, otherwise the width will be returned always in pixels.

With this, the gray background of the sidebar has the proper width, not
e.g. fifteenth of the expected value (using default zoom).

Change-Id: I4380ee0ba6bcda97cf71735161dbdc826e7a2532