sd lok: ccu#1295 force async image swap
This helps ensure that images are swapped in when we actually
render tiles. Previously we'd sometimes have placeholders
instead of the image, which results in either an invalidate
(+rerender of that tile) once the image is swapped in (for
normal tiles) or a permanently missing image in the preview tiles.

Change-Id: I1a16a913faf9fad20e40a5d1aad3de187038c7a2
Reviewed-on: https://gerrit.libreoffice.org/19890
Reviewed-by: Andrzej Hunt <andrzej@ahunt.org>
Tested-by: Andrzej Hunt <andrzej@ahunt.org>