improve compile time SSE2 detection, fix broken android build
Prereq. to enable runtime SSE2 detection is that the compiler
supports it in the first place. MSVS and GCC use different
compiler flags for this so use __LO_SSE2_AVAILABLE__ to make this
build platform independent.

emmintrin.h is unavailable on ARM Android so include this and
compile the SSE2 specific code only when we are sure we can build
SSE2 code (__LO_SSE2_AVAILABLE__ is defined).

Change-Id: I212c4e0b99a314d087b9def822a81325b25f3469