sw lok comments: fix position of blinking cursor after mouse click
LOK always works in absolute twips (origo being the top left corner of
SwEditWin), so not only the callbacks have to translate relative twips
to absolute ones, but the opposite have to be done for mouse event
coordinates.

With this, clicking at a random position inside a comment places the
blinking cursor at a reasonable position, not always at 0,0.

Change-Id: Ic8d20f177acd9e1908acf17698c53a1470bd4aec